Handle distributed mode lock exception. In distributed mode, Orientdb will occasionally return a 500 response with error code com.orientechnologies.orient.server.distributed.task.ODistributedRecordLockedException instead of an MVCC error. This commit also changes the test runner to attempt to detect whether the test orientdb node is running in distributed mode, and a) do slightly different matching for the MVCC test, and b) disable some tests involving db deletion that will always fail in distributed mode, see https://github.com/orientechnologies/orientdb/issues/3746.
This change also necessitated a slight alteration to the error message
regex, as the DistributedRecordLockedException doesn't fit the usual
format of <ERRORCLASS>:<ERRORMESSAGE>.

Fixes #9

+++ spec/orientdb_client_spec.rb
@@ -91,7 +91,7 @@ RSpec.describe OrientdbClient do
         before do
           client.connect(username: username, password: password, db: db)
         end
-        
+
         context 'with valid database parameters' do
           it 'creates a database' do
             client.create_database(temp_db_name, 'plocal', 'document')
@@ -128,55 +128,57 @@ RSpec.describe OrientdbClient do
     end
 
     describe '#delete_database' do
-      before(:each) do
-        if !client.database_exists?(temp_db_name)
-          client.create_database(temp_db_name, 'plocal', 'document', username: valid_username, password: valid_password)
-        end
-      end
-
-      context 'without existing connection' do
-        context 'with valid authentication options' do
-          it 'deletes the database' do
-            client.delete_database(temp_db_name, username: username, password: password)
-            expect(client.database_exists?(temp_db_name)).to be false
+      if !$distributed_mode
+        before(:each) do
+          if !client.database_exists?(temp_db_name)
+            client.create_database(temp_db_name, 'plocal', 'document', username: valid_username, password: valid_password)
           end
         end
 
-        context 'with invalid authentication options' do
-          it 'raises UnauthorizedError' do
-            expect do
-              client.delete_database(temp_db_name, username: 'foo', password: 'bar')
-            end.to raise_exception(OrientdbClient::UnauthorizedError)
+        context 'without existing connection' do
+          context 'with valid authentication options' do
+            it 'deletes the database' do
+              client.delete_database(temp_db_name, username: username, password: password)
+              expect(client.database_exists?(temp_db_name)).to be false
+            end
           end
 
-          it 'does not delete the database' do
-            begin
-              client.delete_database(temp_db_name, username: 'foo', password: 'bar')
-            rescue
-            ensure
-              expect(client.database_exists?(temp_db_name)).to be true
+          context 'with invalid authentication options' do
+            it 'raises UnauthorizedError' do
+              expect do
+                client.delete_database(temp_db_name, username: 'foo', password: 'bar')
+              end.to raise_exception(OrientdbClient::UnauthorizedError)
+            end
+
+            it 'does not delete the database' do
+              begin
+                client.delete_database(temp_db_name, username: 'foo', password: 'bar')
+              rescue
+              ensure
+                expect(client.database_exists?(temp_db_name)).to be true
+              end
             end
           end
         end
-      end
 
-      context 'with existing connection' do
-        before do
-          client.connect(username: username, password: password, db: db)
-        end
-        
-        context 'with valid database parameters' do
-          it 'deletes the database' do
-            client.delete_database(temp_db_name)
-            expect(client.database_exists?(temp_db_name)).to be false
+        context 'with existing connection' do
+          before do
+            client.connect(username: username, password: password, db: db)
           end
-        end
 
-        context 'with no matching database' do
-          it 'raises a ClientError' do
-            expect do
-              client.delete_database(temp_db_name + 'baz')
-            end.to raise_exception(OrientdbClient::ClientError, /OConfigurationException/)
+          context 'with valid database parameters' do
+            it 'deletes the database' do
+              client.delete_database(temp_db_name)
+              expect(client.database_exists?(temp_db_name)).to be false
+            end
+          end
+
+          context 'with no matching database' do
+            it 'raises a ClientError' do
+              expect do
+                client.delete_database(temp_db_name + 'baz')
+              end.to raise_exception(OrientdbClient::ClientError, /OConfigurationException/)
+            end
           end
         end
       end
@@ -674,7 +676,8 @@ RSpec.describe OrientdbClient do
       jim_rid = jim['result'][0]['@rid']
       bob_rid = bob['result'][0]['@rid']
       thrs = []
-      expect do 
+      err = nil
+      begin
         thrs << Thread.new do
           100.times do
             client.command("create edge Friend from #{jim_rid} to #{bob_rid}")
@@ -688,7 +691,16 @@ RSpec.describe OrientdbClient do
           end
         end
         thrs.each { |t| t.join }
-      end.to raise_exception(OrientdbClient::MVCCError, /OConcurrentModificationException/)
+      rescue => e
+        err = e
+      ensure
+        if $distributed_mode
+          correct_error_raised = err.is_a?(OrientdbClient::MVCCError) || err.is_a?(OrientdbClient::DistributedRecordLockedException)
+        else
+          correct_error_raised = err.is_a?(OrientdbClient::MVCCError)
+        end
+        expect(correct_error_raised).to be true
+      end
     end
   end
